refactor: change invalidatePropertyCache argument

+++ models/classes/ClassProperty/RemoveClassPropertyService.php
@@ -22,6 +22,7 @@ declare(strict_types=1);
 
 namespace oat\tao\model\ClassProperty;
 
+use core_kernel_classes_Property;
 use oat\oatbox\event\EventManager;
 use oat\generis\model\OntologyAwareTrait;
 use oat\tao\model\Lists\DataAccess\Repository\ParentPropertyListCachedRepository;
@@ -73,9 +74,7 @@ class RemoveClassPropertyService extends ConfigurableService
                         $index->delete(true);
                     }
 
-                    if ($property->getRange()) {
-                        $this->invalidatePropertyCache($parsedBody['uri'], $property->getRange()->getUri());
-                    }
+                    $this->invalidatePropertyCache($property);
 
                     return true;
                 }
@@ -85,14 +84,16 @@ class RemoveClassPropertyService extends ConfigurableService
         return false;
     }
 
-    private function invalidatePropertyCache(string $propertyUri, string $listUri): void
+    private function invalidatePropertyCache(core_kernel_classes_Property $property): void
     {
-        $this->getParentPropertyListCachedRepository()->deleteCache(
-            [
-                'propertyUri' => $propertyUri,
-                'listUri' => $listUri
-            ]
-        );
+        if ($property->getRange()) {
+            $this->getParentPropertyListCachedRepository()->deleteCache(
+                [
+                    'propertyUri' => $property->getUri(),
+                    'listUri' => $property->getRange()->getUri()
+                ]
+            );
+        }
     }
 
     private function getEventManager(): EventManager
